+++ src/constants/contentTypes.ts
@@ -136,7 +136,7 @@ export const CT_DOMAIN = 'application/vnd.sap.adt.domains.v2+xml';
 
 // Data Elements
 export const ACCEPT_DATA_ELEMENT =
-  'application/vnd.sap.adt.dataelements.v2+xml, application/vnd.sap.adt.dataelements.v1+xml';
+  'application/vnd.sap.adt.dataelements.v1+xml, application/vnd.sap.adt.dataelements.v2+xml';
 export const CT_DATA_ELEMENT = 'application/vnd.sap.adt.dataelements.v2+xml';
 
 // Structures